dunno why the spec had this a float when an int will do

+++ chess_player/Scorer.py
@@ -7,7 +7,7 @@ class Scorer(ABC):
         pass
 
     @abstractmethod
-    def evaluate(self, board: chess.Board) -> float:
+    def evaluate(self, board: chess.Board) -> int:
         pass
